refactor(orchestration): split definitions.py into per-domain modules
Extract 447-line monolithic `definitions.py` into focused modules:

- `_factories.py` — shared primitives (SQLMesh translator, dlt helpers,
  `sqlmesh_project` multi-asset, Soda check factory, freshness policies)
- `domains/analytics.py` — cross-domain marts + Soda checks
- `domains/ebird.py` — dlt ingestion + marts + checks + job + schedule
- `domains/noaa.py`  — same shape
- `domains/usgs.py`  — same shape
- `definitions.py`   — 47-line assembly only

Each domain module is self-contained: imports only from `_factories`,
owns its own `dlt_asset_keys`, `sqlmesh_asset_keys`, `asset_checks`,
`daily_pipeline`, and `schedule`. No cross-domain imports.

Ticket: ticket:definitions-split (Phase 1, initiative:scaffold-polish).
Validation: 30 tests pass, ruff + ruff-format clean, mypy error count
dropped from 10 to 5 (5 remaining are pre-existing module-stub noise),
`Definitions` resolves 45 assets / 4 jobs / 3 schedules / 17 checks,
mkdocs --strict clean.
